requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
inflate64==1.0.0
loguru==0.7.2
multidict==6.0.4
multivolumefile==0.2.3
orjson==3.9.10
psutil==5.9.6
py-cord==2.4.1
py7zr==0.20.8
pybcj==1.0.2
pycares==4.4.0
pycparser==2.21
pycryptodomex==3.19.0
pyi18n-v2==1.2.1
pyppmd==1.1.0
python-decouple==3.8
python-magic==0.4.27
PyYAML==6.0.1
pyzstd==0.15.9
rarfile==4.1
texttable==1.7.0
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
